Move global chat to components folder, update imports
