--- src/global-chat/Message.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import './Message.css';
-
-function Message({username, message}) {
-    return (
-        <div className="message_view">
-            <p><strong>{message.username}</strong>: {message.text}</p>
-        </div>
-    )
-};
-
-export default Message;
