--- src/global-chat/global-chat.css
@@ -1,48 +1,0 @@
-.chat-form {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.chat input {
-  width: 100%;
-  box-shadow: rgba(50, 50, 93, 0.25) 0px 6px 12px -2px,
-    rgba(0, 0, 0, 0.3) 0px 3px 7px -3px;
-  padding: 10px 5px;
-  font-family: inherit;
-  background-color: #f5f4f6;
-  border: none;
-  transition: all 0.2s;
-}
-
-.chat input:focus {
-  box-shadow: 0 5px 10px rgb(0 0 0 / 50%);
-}
-
-.chat {
-  box-shadow: rgba(50, 50, 93, 0.25) 0px 6px 12px -2px,
-    rgba(0, 0, 0, 0.3) 0px 3px 7px -3px;
-  font-size: 26px;
-  line-height: 35px;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  background-color: #f5f4f6;
-  width: 100%;
-  min-width: 80vw;
-  grid-row: 3/4;
-}
-
-.chat_userinfo {
-  margin: 10px 0 0;
-  word-break: break-word;
-}
-
-.chat_messages {
-  height: 150px;
-  word-break: break-word;
-  overflow-y: scroll;
-  box-shadow: rgba(50, 50, 93, 0.25) 0px 6px 12px -2px,
-    rgba(0, 0, 0, 0.3) 0px 3px 7px -3px;
-}
